src/api/requirements.txt
azure-search-documents==11.4.0
fastapi==0.110.0
langchain==0.1.13
langchain-openai==0.1.1
loguru==0.7.2
pydantic-settings==2.2.1
uvicorn==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9
openai==1.14.3
tiktoken==0.6.0
azure-identity==1.15.0
azure-cosmos==4.6.0
unstructured==0.13.3
langchain-openai==0.1.3
